feat(app.ts): add logic to create an alert if a knife is detected in the bus feat(prisma/schema.prisma): add optional 'objectKey' field to the Alert model
The changes in app.ts add a conditional check to detect if the label "Knife" is present in the detectedLabels array. If it is, a new alert is created using the Prisma ORM. The alert has a type of "PERICULOUS_OBJECT", an expiration time of 2 minutes from the current time, and a description stating that a knife was detected in the bus.

The changes in prisma/schema.prisma add an optional 'objectKey' field to the Alert model. This field can be used to store additional information about the detected object, if needed.